backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
